Add image-c9s.yaml
Adds a dedicated image config for CentOS Stream 9.

`oc adm release new` and the CVO require the `machine-os` label to be present
on one of the payload images.
For c9s, we let COSA add the required labels to the ostree container image.

In OKD/SCOS, this image is aliased as both `machine-os-content`
and `centos-stream-coreos-9` in the payload.