Bump to OVS 3.1 on RHEL9
Container side bumps to 3.1 in https://github.com/openshift/ovn-kubernetes/pull/1544
and we should to keep the host OS side in sync.

This reverts part of https://github.com/openshift/os/pull/1140
since (currently) OVS 3.1 is cross-tagged into server-ose
repos since it's not in FDP quite yet.

When FDP picks 3.1 up in a couple weeks we'll want to un-revert
this hunk and somehow pull from either of [FDP, server-ose] to
get whichever has the newest RPMs.