Revert "manifest/9.2: Ensure package come from RHEL by default"
We already have to use coreos-installer from the RHAOS repo and it's
likely that we will have to do it again in the future for 4.14.

Remove this to avoid having to make a PR for each package.

This reverts commit 49f9824db2b6f929567d2d41cea6c97df766cf37.

(cherry picked from commit e346f8a350269e3966db9ade9df694e971cd3dfe)