overlay: add 30rhcos-nvme-compat-udev
The systemd in RHEL 9.2 now ships this commit:
https://github.com/systemd/systemd/commit/5118e8e71d

This changes how symlinks created from device serials with leading spaces are interpreted, from e.g.

```
nvme-QEMU_NVMe_Ctrl__foobar -> ../../nvme0n1
```

to

```
nvme-QEMU_NVMe_Ctrl_______foobar -> ../../nvme0n1
```

This in turn breaks anything/anyone that relied on the previous
symlinks. Things in the storage stack for example that auto-discover
block devices, or simply just customers customizing their storage.

Add back a rule similar to the previous version so that we keep creating
the previous symlink. Also ship this in the initramfs in case Ignition
configs referenced them.

The motivating factor for the systemd change was to fix incorrect
behaviour when e.g. the model name contained a `/`; udevd would create
a literal directory as a result:

https://github.com/systemd/systemd/issues/19309

This rule we're adding back is still susceptible to this issue, though
we haven't seen it in the wild for RHCOS. We could deviate from the
old rule and have something custom to handle this but it's probably
not worth the effort. The new rule is still there creating the fixed
symlink, so users could still use that one if such a case arises.

Fixes: https://issues.redhat.com/browse/OCPBUGS-11375
See also: https://github.com/systemd/systemd/issues/27155