OCPBUGS-11124, OCPBUGS-11411: overlay: Mask pcrphase service
This PR masks systemd-pcrphase service which is not used at the moment.
Its existence creates issues as it depends on the remote-fs and as a
consequence blocks access to the system if network configuration is not
correct.

Anyway in OpenShift we are not using remote home directories, so this
would not be useful for us.

We are also modifying dependencies of systemd-user-sessions.service so
that we are explicitly saying that we do not need network in order to
allow access to the node.

Fixes: OCPBUGS-11124
Fixes: OCPBUGS-11411