tests/trust-cpu: support v6.x kernels
RHEL 10 will have a v6.x kernel (and RHEL 9.6 and onwards as an option).
We'd like to do some early testing of RHCOS and OCP with that kernel
since it also includes other major changes such as a new RPM name
(`kernel-redhat`).

Adapt this test to run fine on 6.x kernels, automatically noop'ing on
versions where we know `CONFIG_RANDOM_TRUST_CPU` was removed as per
https://github.com/coreos/fedora-coreos-tracker/issues/1369.

While we're here, use `architectures` to gate the whole test on x86_64
rather than noop'ing at runtime.