Move non-C9S/RHEL packages to dedicated manifest
The packages moved to a new manifest with this change are required by
OpenShift/OKD but are not present in CentOS Stream and RHEL repos.
Versioning of these packages is coordinated and tested for each minor
OpenShift release.

In the future, this manifest could be excluded from base SCOS composes
and the manifest be layered on top in a derivation step, i.e.
different layers corresponding to specific OpenShift minor versions
could be layered onto same shared minimal SCOS base.