overlay: stop disabling `iscsid.socket`
The iSCSI daemon is now socket-activated so that it's only running when
needed rather than always enabled. We're breaking that by disabling
`iscsid.socket`.

This effectively reverts 929ac48 ("c9s: Disable iscsid.socket"). It's
not certain why this was done, but it was likely to workaround failing
tests. These tests should be fixed now[[1]] so we should be able to stop
doing this.

[1]: https://github.com/coreos/coreos-assembler/pull/3275

(cherry picked from commit 8c9427029cd6da75756a797f04d85c51caebd4ec)