drop hardcoded openvswitch user/group definitions
The RPM is now using systemd-sysusers fragments [1] so we can drop the
hardcoded definitions. One problem here, though, is that the hugetlbfs
group never gets added to the openswitch user [2] so let's add a
workaround for that for now.

One side effect of this is that it does change the previously defined
UID/GIDs from 800/801 to different values (dynamically generated at build
time). I assume this is OK because of some of the discussion in [1].

[1] https://github.com/openshift/os/issues/1274#issuecomment-1593690794
[2] https://github.com/openshift/os/issues/1274#issuecomment-1595860275