OCPBUGS-15843: Enable rhcos-fips-finish.service in diskless cases
Allow the user to enable FIPS by passing fips=1 on the kernel command
line, even in diskless cases such as the live ISO.

The rhcos-fips.service (which reads the MachineConfig and sets up fips=1
on the kernel command line before triggering a reboot) works only once
the OS is written to disk. The same is not true of
rhcos-fips-finish.service, which runs after fips=1 is passed at boot
time. The former is not needed if fips=1 is already passed on the kernel
command line, but the latter is required to set up the crypto policies.