Re-enable `ostree-container-inject-openshift-cvo-labels` knob
This was previously enabled (#1048) and then disabled again (#1084)
because `oc` doesn't know how to handle multiple images with those
labels in the release payload. We'll need to solve this eventually if
we want to be able to ship multiple OS images in the payload (that's
tracked in #1047), but we don't need to block on this if we can remove
the legacy `machine-os-content` at the same time.

See also: https://github.com/openshift/driver-toolkit/issues/101
See also: https://github.com/openshift/machine-config-operator/pull/3364