manifest-rhel-9.4: fake out RHEL_VERSION
During this transitionary period, we're inheriting from the C9S
release package which doesn't define `RHEL_VERSION`. Since we're trying
here to pretend to be a normal RHEL-based RHCOS, let's have a valid
`RHEL_VERSION` string too.

Fixing this here means that consumers of this file don't need to worry
about this weird case where RHCOS is pretending to be RHEL but is really
CentOS Stream.

Fixes: #1384