catch coreos.ignition.failure err on 4.16 pipeline.
After failing to reproduce `coreos.ignition.failure` locally I am removing it from denylist in order to catch errors in the pipeline. The goal of this PR is to allow it in RHCOS 4.16, catch the errors from the log and re-introduce `coreos.ignition.failure` to kola-denylist while working on solution.

See: https://github.com/coreos/coreos-assembler/issues/3670