manifests: default to rhel-9.4 variant
This mirrors the ART switchover to defaulting to the 9.4 builds in CI
and nightlies.

Note the pipecfgs already explicitly specify variants for both the
4.16-9.2 and 4.16-9.4 variants. So this change is mostly symbolic (pun
intended), but does affect the default variant built by developers.

While we're here, do some other subtle cleanups. I think it was
confusing how e.g. `image-rhel-9.4.yaml` symlinked back to
`image-rhel-coreos-9.yaml`, like it was going back up one
abstraction layer. Instead, I think it's clearer to have the concrete
implementations symlinks among each other if they happen to be
identical. We do this now with `image-rhel-9.2.yaml` being a
symlink to its 9.4 counterpart.

git makes it seem like it's a larger change than it is because of the
type changes. But it's basically just some renames and symlinking. No
content was changed.