Move `/etc/motd` generation to `packages-openshift.yaml`
The MOTD message is strongly associated with OpenShift, so let's make it
part of `packages-openshift.yaml`.

Now, to do this in practice is actually quite cumbersome. To keep the
MOTD wording similar to what we have and spell out the variant name,
we need to have the postprocessing that edits `/etc/os-release` happen
first. And to do that... we're forced to factor out the script that
does that to a separate manifest so that we can include it *after*
`packages-openshift.yaml`.

This is all pretty messy, but is also temporary; once we execute the OCP
split, we will remove those extra manifests. (See next commit which does
this for the c9s variant.)