manifest-rhel-9.4: pin to NetworkManager-1.46.0-1.el9
The latest NM currently in c9s is hitting issues. But anyway, that NM is
not going to end up in 9.4. So let's just temporarily pin to the latest
version that is.

We should be able to clean this up once the rhel-9.4 variant stops
consuming from c9s.

See also: https://issues.redhat.com/browse/RHEL-29856