manifest-c9s.yaml: name centos-stream-release, not centos-release
The latter was replaced by the former, which now Provides
`centos-release`.

This also fixes the fact that `ci/prow-entrypoint.sh` queries these
manifests and checks for `centos-stream-release` in the list of
packages. The querying happens directly on the flattened manifest,
without any depsolving, so the fact that `centos-release` is provided by
`centos-stream-release` is not relevant at all.

It would have also worked to keep `centos-release` here and keep
querying `centos-release` in the script, but it seems cleaner anyway
to be canonical and list the actual package name rather than one of its
virtual Provides.