Manifests: disable default ZRAM config
In #2937 we moved zram-generator to the system-configuration.yaml
manifest so it's shared with RHCOS.
However in the centOS stream 9 zram-generator-0.3.2 [1] ships a
default config file, resulting in a swap device being created
by default.
This ends up breaking kubelet, until kubernetes 1.30
is released. Removing the default config from the package until
c10s

[1] https://gitlab.com/redhat/centos-stream/rpms/rust-zram-generator/-/blob/c9s/zram-generator.conf?ref_type=heads