Remove all RHEL 9.2 references
At this point, we're pretty committed to RHEL 9.4 and it's very unlikely
that we'll revert back to RHEL 9.2. If we do, we can just revert this
PR. This variant isn't even building right now in the pipeline so
there's no attention paid to it and no point in carrying it.