c9s: always get containers-common from appstream
There's a messy situation right now where the containers-common package
is higher versioned in OCP than in c9s proper. And because we need the
OCP repo for now to compose SCOS, we get the OCP one, which causes
issues because unlike the c9s version, it doesn't ship the RHEL keys.

Work around this by pinning containers-common to the c9s-appstream repo.

While we're here, improve error-handling so that we output stderr if
podman fails.

See also: https://github.com/openshift/os/issues/1505#issuecomment-2102988693

Fixes: https://github.com/openshift/os/issues/1505