kola-denylist: add ext.config.shared.networking.* on c9s
A bunch of tests are failing because `systemd-network-generator.service`
is hitting SELinux denials:

https://issues.redhat.com/browse/RHEL-47033

Some of these tests are testing things that don't strictly need that
service, but test kargs that trigger it anyway. Other tests do (e.g.
`ifname=`) and so that functionality is indeed currently broken on c9s.

That said, there's continuing requests to have preliminary c9s
bootimages for testing. So let's not block on this issue for now.