common.yaml: add subscription-manager
We currently ship dnf in RHCOS but it's cumbersome to use because the
librhsm magic we had for rpm-ostree to make content available during
builds on a subscribed node doesn't work with dnf. For that, it seems
like we need full-blown subscription-manager in the container image.

Add it.

This matches centos-bootc/rhel-bootc shipping subman as well, though I'd
like in the future to investigate lowering the dependency requirements
to just a dnf plugin + librhsm.

Fixes: https://issues.redhat.com/browse/OCPBUGS-35247