extensions/Dockerfile: fix for okd-c9s variant
In fc9e703 ("extensions/Dockerfile: install SIG GPG keys on CentOS"), we
fixed the extensions container build for the `c9s` variant, but that
broke it for the `okd-c9s` variant, which CI currently still tests.

We should switch CI over to target the `c9s` variant and get rid of
`okd-c9s`, but for now at least let's keep it happy.

The problem is that in the `okd-c9s` variant, we disable all the repos
as part of `packages-openshift.yaml`. But anyway, we also already
include the various release packages containing the GPG keys there, so
we can skip this step entirely.

Fixes fc9e703 ("extensions/Dockerfile: install SIG GPG keys on CentOS").
Fixes: #1564