extensions/Dockerfile: use latest Fedora to generate repo
In this stage of the build, we're just using Fedora to generate
repodata. Since this is a pretty basic operation, let's just always use
the latest Fedora instead of hardcoding a specific version. This ensures
that e.g. we're never going to use an EOL Fedora release.

This became an issue after 8301c67 ("extensions/Dockerfile: use the
FCOS defined fedora.repo to set up container"), which was backported
to older RHCOS branches that used EOL Fedora releases. Before, we were
using the mirrors, which papered over the fact that the content for EOL
releases actually moved to another location. But now that we're using
the canonical Fedora server, we're directly exposed to that.

Really, we should just stop hardcoding Fedora versions in this repo. We
shouldn't be using EOL releases at all as part of our builds or tests.
(Obviously a massive offender here is cosa... though that's a discussion
of its own.)

See also: https://github.com/openshift/os/issues/1546