rhel-9.4/c9s: make versioning scheme semver compatible
There's a bunch of things out there that currently parse the RHCOS
version string. The semantics of this string will change as part of this
change, but we should at least try to retain the same semver-compatible
format.

E.g. currently, an RHCOS version string of the `ocp-rhel-9.4` variant
is 418.94.202409090849-0. Before this PR, An RHCOS version string
of the `rhel-9.4` variant is 9.4.202409090849.0, which is not semver
compatible. With this PR, the latter would now be 9.4.202409090849-0
(note the change from `.0` to `-0`).