packages-openshift: add compat tmpfiles.d dropin
We can't rely on rpm-ostree's var-to-tmpfiles.d conversion anymore
for layered packages. Let's not worry about it for new packages in the
future, but at least for the existing ones, let's add a compat dropin
that adds back what we lost.

Though I could also imagine exposing a e.g.
`rpm-ostree generate-tmpfiles.d` that we could use here, or just lift
that code into a separate utility/SDK since it's globally useful.