common.yaml: disable composefs on ppc64le for now
There are issues with composefs on ppc64le right now:
https://issues.redhat.com/browse/RHEL-70199

Work around this for now by disabling composefs on ppc64le only.

We don't want to ship to production with this kind of delta across
arches, but at least for now this allows us to keep running tests on
arches where it does work.

Note this will also run on 9.4 variants, but it's harmless there since
`prepare-root.conf` doesn't exist.