tests/replace-rt-kernel: ensure we always pick a different EVR
Right now, only on c9s-based systems, we have logic to ensure we pick a
different EVR. But even in the RHEL case, we need this logic because we
might be in the period where c9s and RHEL are in sync and so the kernel
builds have the same EVRs in both RHEL and c9s.

Work around this by just always doing the repoquery and filtering.

This is technically an rpm-ostree bug I think, but highly unlikely to
happen in the wild because you normally replace with a kernel from the
same vendor.