el-only variants: drop hours and minutes from version timestamp
As part of the layered node image work, I was initially thinking we'd
still keep the e.g. `419.96...` format there, but I'm now thinking we
_don't_ override the versioning info from the base RHCOS layer anymore.
So e.g. `bootc status` on an OCP node would still show `9.6...`. This is
potentially breaking; I'm ready to help fix any issues that occur from
this change in any OCP code that parses the version string today.

So then, it becomes much more enticing to also bundle with this change
the suggestion in #1617 since any consumers  will have to adapt to the
new version string anyway.

So let's do that.

Closes: #1617