el-only variants: drop pungi-related comments
It was an aspiration at the beginning of this effort to try to think of
these streams as kind of follow-on jobs to pungi composes that happen
in CentOS Stream and RHEL. And so I had an idea of matching exactly the
Pungi compose ID we were composing from to reinforce that.

In practice, it was quite awkward to query compose IDs and also do it in
a race-free way. It didn't seem worth the added complexity, so I backed
it out. It's also currently not clear what the future of Pungi is.

So let's just remove the comments related to it.