common.yaml: disable composefs on all arches on el9
There are issues currently on ppc64le and kernel-64k in el9 that break
with composefs. So it's not sufficient to do an architecture check
anyway. See: https://issues.redhat.com/browse/RHEL-63985.

The upcoming release of libostree will turn on composefs by default, so
it's not sufficient to delete `prepare-root.conf`; we need to explicitly
turn it off. See: https://github.com/ostreedev/ostree/pull/3353.