ci/get-ocp-repo.sh: add `--output-dir`
In the node layering case, we're directly bind-mounting the build
context dir. It's mounted read-only, so we can't write `ocp.repo` there.
And even if we could, we shouldn't modify the build context if we don't
need to.

Add a new `--output-dir` option to allow changing the directory to which
to write `ocp.repo`. Use this in the node layering flow to output to
`/run/yum.repos.d`, which is an already established API for transient
repo definitions.