manifest-ocp-rhel-9.6: get toolbox from ART plashet
RHCOS is trying to move to a point where all container tool
packages come from RHEL instead of the ART plashet[0]. This is part
of the effort to have a RHEL only bootstrap image and a layered container
image for the node. Unfortunately there is already a package in RHEL
called toolbox which conflicts with the RHCOS toolbox.

Ensure that RHCOS gets the RHCOS toolbox by only getting the package
from the ART plashet.

[0] https://issues.redhat.com/browse/RHELPLAN-170883