Containerfile: use secrets API for yum repo injection
Instead of the `/run/yum.repos.d` mechanism, use the secrets API which
is already well-known and does what we want here, which is to avoid
having to copy repo files into the image.

The `RUN --mount` option gracefully ignores the case where no secret was
provided, which is also the semantic we want for this.