Containerfile: add metadata in last layer of node image
Prior to the layered RHCOS enhancement[[1]], there was tooling in
ART and in the RHCOS release browser (notably linked from the release
controller) that relied on the commit metadata JSON file published
separately to be able to inspect the RPM contents of a build.

With the move to a layered node image, we lose this. The more
container-native analogue would be SBOMs, which we should get for free
once we move to Konflux. However, until then, let's provide an easy-to-
generate alternative that still allows avoiding to download the whole
image.

We add a new semantic where the last layer of the node image is actually
a metadata-only layer. Then, tooling like `oc image extract` capable of
extracting only specific layers from an image can be used to pull it
down. E.g.

```
$ oc image extract quay.io/jlebon/rhcos:latest[-1] --file /usr/share/openshift/base/meta.json
$ jq . meta.json
```

Add a label to signal that this image implements this semantic.

Compared to SBOMs, this has the _major_ advantage that it's in-band
and implicitly gets preserved as the image is slung around (whereas an
attached SBOM may not until the tooling is careful to do so).

But like SBOMs, it has the pitfall that the metadata is only valid for
the original node image, and obviously skews once more layering is done.
The source of truth for the package list is always the rpmdb.

That said, a client implicitly knows if the metadata no longer matches
the node content simply by the fact that it wouldn't be in the last
layer.

[1]: https://github.com/openshift/enhancements/blob/master/enhancements/rhcos/split-rhcos-into-layers.md