Move ose-aws-ecr-image-credential-provider to CoreOS base
The ose-aws-ecr-image-credential-provider package is used in AWS by the
kubelet to get ECR credentials, as described by:

https://kubernetes.io/docs/tasks/administer-cluster/kubelet-credential-provider/

However, it's also currently used at least by ROSA to provide a pull
secret to podman during provisioning. This use case then broke when we
switched to el-only bootimages and stopped including it.

The long-term goal is to have that functionality in RHEL proper, but at
least for 4.19 let's work around this by lowering the package down to
the CoreOS layer.

See also: https://issues.redhat.com/browse/RHEL-82921