ci/get-ocp-repo.sh: pull in mirror repos when building scos
Recently, building scos extensions hot this error:
```
error: Packages not found: kernel-rt-core-5.14.0-570.el9, kernel-rt-devel-5.14.0-570.el9, kernel-rt-kvm-5.14.0-570.el9, kernel-rt-modules-5.14.0-570.el9, kernel-rt-modules-extra-5.14.0-570.el9
```

This is because the baseos image was built with the 570 version of the kernel, but the composes had already moved on to the 571 version. While pulling the mirror repos fixes this for now, it would be ideal to always
build from mirror repos as composes seem to have a faster cadence. We should probably also have the node iamge building be triggered at the same time the base images are being built to avoid these races.