Move ose-aws-ecr-image-credential-provider to rhel-9.6 only
Limit the hack for this to RHEL 9.6 only for now by lifting it up to
the 9.6-specific manifest. Otherwise, this breaks c9s builds which don't
have access to the OCP plashets.

Also, just keep it listed in `packages-openshift.yaml`. In RHCOS node
image builds, this is a no-op since it'll already be in the base. In
SCOS node image builds, this will layer it.