packages-openshift.yaml: bump OVS version to 3.3.
Open vSwitch 3.1 is EoL upstream and so the support of this version
in FDP is limited.  To ensure timely bug fix delivery for OCP 4.14+ we
need to upgrade to the current LTS stream, which is Open vSwitch 3.3.
It will be fully supported for much longer.

Consolidation of versions used across releases of OpenShift will also
reduce pressure on QE teams allowing to drop older releases not used
by any products.

ovn-kubernetes container image will follow the lead and move to this
version as well.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>