packages-openshift: define repos list there
Previously, when building the layered node image, we were relying on the
default repo enablement settings. This though is at the source of a lot
of complexity because then we need to make sure that we only inject just
the repos that we need with the right enablement. See e.g. the complex
logic in `get-ocp-repo.sh`.

Let's instead match the semantics already in use by the base compose and
extensions builds, both of which explicitly list the repos to enable.
This means that we can be a lot less careful in what repo definitions we
inject into the build environment, knowing only the necessary ones will
be enabled.

This is pretty easy to do now that (1) rpm-ostree suppports inlined
treefiles, and (2) `treefile-apply` supports a `--var` option to define
variables at invocation time.