extensions: make rhel-9.4-appstream global
There's a subtlety here that's causing issues. Any repo that's available
on all arches needs to be listed in the global repos list. Otherwise,
we may have skew across arches for packages not even related to the
extensions for which the repo was added.

See also: https://issues.redhat.com/browse/OCPBUGS-52293