okd: add filtered repo for 4.19 plashet
We only want certain packages to come from the 4.19 plashet. And we
can't just rely on NVRs because the plashet may sometimes win. Long-term
we should sever that dependence on ART packages, but for now, let's add
a hack to essentially generate a repo on the fly from the 4.19 repo with
the filters we need.

The advantage of doing it this way instead of e.g. in the
`get-ocp-repo.sh` script is that this applies both in CI and locally.