build-node-image.sh: include yum repos from git as part of build
When building the node image, we want it to be able to use the repos
defined in the git repo itself. For the local developer case, this is
implicitly done by `cosa podman-build`:

https://github.com/coreos/coreos-assembler/blob/325ca2be9fc349ba329f49fab65ea207ba338d19/src/cmd-podman-build#L34

But nothing does that in the OpenShift CI case.

So do it.

We should then be able to delete that line from `cosa podman-build` to
avoid duplicate definitions.

Note that in the CentOS Stream case, the canonical repos live in the
node image already so we could use that, but it's cleaner I think to
ensure we're consistently using the same repo definition files whether
we're building the base image, the node image, or the extensions (e.g.
the repo IDs are different, and there are subtleties between using the
compose vs mirror repos).

While we're here, also make sure that we delete the `okd.repo` file we
injected; we don't want that in the final image.