OCPBUGS-54594: update bootloader on aarch64 systems
The aarch64 kernel changed the file format [1] [2] and older
RHEL8 based systems (4.12 and 4.11) need to update the bootloader
otherwise the system won't boot when they get upgraded to 4.19
based on RHEL 9.6.

Let's add a systemd unit here that will update the bootloader.
Also need to add code that will handle the RAID case because
bootupd doesn't currently handle that case.

It's worth mentioning that we did hit this upstream in Fedora
CoreOS as well [3] and we took this fix from that [4] and another
similar issue [5] as inspiration here.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2162369
[2] https://issues.redhat.com/browse/RHEL-25537
[3] https://github.com/coreos/fedora-coreos-tracker/issues/1441
[4] https://github.com/coreos/fedora-coreos-config/pull/2308
[5] https://github.com/coreos/fedora-coreos-config/pull/3042

Fixes: https://issues.redhat.com/browse/OCPBUGS-54594