docs: delete obsolete docs
All these files are obsolete. Delete them.

Keep the FAQ here for now even though most of it pertains to the base
compose. It's been heavily linked and it's a shame to break those links.
If we want to move it to the rhel-coreos-config repo, we could replace
the one here with a link to the new location.

Keep the vSphere screenshot since it's referenced from the FAQ.