denylist: add 2 tests under ext.config.var-mount
These tests started failing but i cannot replicate the failure so let's
denylist these until we investigate.

The last known good build is 412.86.202504240545-0
I tried to pin the kernel to 4.18.0-372.144.1.el8_6 because the only
with the previous good build is the kernel updating to 4.18.0-372.146.1.el8_6.
However the tests still fail.
With a local build using the same COSA image as the pipeline I am not
seeing this tests fail, so it must be something with the pipeline
environnement.