denylist: add `rhcos.network.init-interfaces-test` entry
This test has been failing since we started running openshift-specific
kola tests in the `build-node-image` job. Let's denylist it for now so
we can unblock the pipeline while we investigate
https://github.com/openshift/os/issues/1852

Also remove the stream definitions from the
`ext.config.version.rhaos-pkgs-match-openshift` etnry. An entry has been
added to the `release-4.19` branch to properly denylist the test for
4.19-9.6. See: https://github.com/openshift/os/pull/1853